out: Fixing the error on CLI when maven settings file doesn't exist
The current version was broken on CLI as the resource was looking for a maven settings file.
This commit is fixing it by adding an IF condition before trying to guess its path.

I also found a problem with PR identification in the decorate_pr attribute.
At the top of the resource, there is a `cd ${project_path}`
Then all conditions are checking for `${project_path}/.git`. The problem is that with the cd occuring above, it will never find this project path as we are already in.

There were three options:
1/ use the absolute path in the configuration of the resource
2/ remove the cd project_path at the top of the resource
3/ prefix the project path with the absolute path to make all conditions passing correctly

I went with the third option that was the less modifying/destroying the logic of the resource